refactor: optimize naming, codepoint handling, and Java file generation logic
- Simplify function calls and redundant logic in JavaTypeName handling.
- Refactor CodePoint implementation to use UInt, add character category determination, and improve logic clarity for Java identifier validation.
- Standardize newline emissions in code writers for Java and common code.
- Enhance ClassName and PackageName handling with new parsing utilities for better organization.
- Add ClassNameTest for canonical name and package name tests.
- Update Java file generation methods to utilize improved name handling and structuring.

+++ codegentle-common/src/jvmMain/kotlin/love/forte/codegentle/common/codepoint/CodePoint.jvm.kt
@@ -4,22 +4,22 @@ import kotlin.text.codePointAt as codePointAtKt
 
 @InternalCodePointApi
 public actual fun String.codePointAt(index: Int): CodePoint =
-    CodePoint(codePointAtKt(index))
+    CodePoint(codePointAtKt(index).toUInt())
 
 @InternalCodePointApi
 public actual fun CodePoint.isLowerCase(): Boolean =
-    Character.isLowerCase(code)
+    Character.isLowerCase(code.toInt())
 
 @InternalCodePointApi
 public actual fun CodePoint.isUpperCase(): Boolean =
-    Character.isUpperCase(code)
+    Character.isUpperCase(code.toInt())
 
 @InternalCodePointApi
 public actual fun CodePoint.charCount(): Int {
-    return Character.charCount(code)
+    return Character.charCount(code.toInt())
 }
 
 @InternalCodePointApi
 public actual fun StringBuilder.appendCodePoint(codePoint: CodePoint): StringBuilder {
-    return appendCodePoint(codePoint.code)
+    return appendCodePoint(codePoint.code.toInt())
 }
