+++ codegentle-java/src/jvmMain/kotlin/love/forte/codegentle/java/codepoint/CodePoint.jvm.kt
@@ -4,7 +4,7 @@ import love.forte.codegentle.common.codepoint.CodePoint
 
 
 internal actual fun CodePoint.isJavaIdentifierStart(): Boolean =
-    Character.isJavaIdentifierStart(code)
+    Character.isJavaIdentifierStart(code.toInt())
 
 internal actual fun CodePoint.isJavaIdentifierPart(): Boolean =
-    Character.isJavaIdentifierPart(code)
+    Character.isJavaIdentifierPart(code.toInt())
